Revert "Clear always-on SRAM area at reset in lc823450.cfg"
This reverts commit 9e5f77ed378f3393038fe792c0bb504b70b22e91.

With this patch, we found that there is a side effect of clearing
SRAM when we load a program into both SRAM and SPI-Flash areas.

However, clearing always-on SRAM area is necessary when debugging
startup code with -rtos nuttx, because the SRAM area might be
inconsistent with the current configuration. To avoid this, you
need to add 'monitor mww 0x02040000 0x0 131072' to .gdbinit
before loading a program.

Signed-off-by: Masayuki Ishikawa <Masayuki.Ishikawa@jp.sony.com>
Tested-by: jenkins